Refactor: 불필요한 매개변수를 받지 않도록 수정함
