+++ src/main/java/com/growup/pms/auth/service/JwtLoginService.java
@@ -2,7 +2,8 @@ package com.growup.pms.auth.service;
 
 import com.growup.pms.auth.domain.SecurityUser;
 import com.growup.pms.auth.dto.LoginRequest;
-import com.growup.pms.auth.dto.TokenDto;
+import com.growup.pms.common.security.jwt.JwtTokenProvider;
+import com.growup.pms.common.security.jwt.dto.TokenDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
@@ -13,12 +14,15 @@ import org.springframework.stereotype.Service;
 @RequiredArgsConstructor
 public class JwtLoginService {
     private final AuthenticationManagerBuilder authenticationManagerBuilder;
-    private final JwtTokenService jwtTokenService;
+    private final JwtTokenProvider tokenProvider;
+    private final RefreshTokenService refreshTokenService;
 
     public TokenDto authenticateUser(LoginRequest request) {
         var authenticationToken = new UsernamePasswordAuthenticationToken(request.getEmail(), request.getPassword());
-        Authentication authentication = authenticationManagerBuilder.getObject()
-                .authenticate(authenticationToken);
-        return jwtTokenService.generateJwtTokens(((SecurityUser) authentication.getPrincipal()).getId(), authentication);
+        Authentication authentication = authenticationManagerBuilder.getObject().authenticate(authenticationToken);
+        SecurityUser principal = (SecurityUser) authentication.getPrincipal();
+        TokenDto newToken = tokenProvider.generateToken(principal);
+        refreshTokenService.renewRefreshToken(principal.getId(), newToken.getRefreshToken());
+        return newToken;
     }
 }
