+++ src/main/java/com/growup/pms/auth/service/JwtTokenService.java
@@ -1,11 +1,11 @@
 package com.growup.pms.auth.service;
 
-import com.growup.pms.auth.dto.TokenDto;
+import com.growup.pms.auth.domain.SecurityUser;
 import com.growup.pms.common.exception.code.ErrorCode;
 import com.growup.pms.common.exception.exceptions.AuthenticationException;
 import com.growup.pms.common.security.jwt.JwtTokenProvider;
+import com.growup.pms.common.security.jwt.dto.TokenDto;
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -14,18 +14,11 @@ public class JwtTokenService {
     private final JwtTokenProvider tokenProvider;
     private final RefreshTokenService refreshTokenService;
 
-    public TokenDto generateJwtTokens(Long userId, Authentication authentication) {
-        return TokenDto.builder()
-                .accessToken(tokenProvider.createAccessToken(userId, authentication))
-                .refreshToken(tokenProvider.createRefreshToken(userId, authentication))
-                .build();
-    }
-
-    public TokenDto refreshJwtTokens(String refreshToken) {
-        validateRefreshToken(refreshToken);
-        Long userId = tokenProvider.getUserIdFromToken(refreshToken);
-        refreshTokenService.updateRefreshToken(userId, refreshToken);
-        return generateJwtTokens(userId, tokenProvider.getAuthentication(refreshToken));
+    public TokenDto refreshJwtTokens(SecurityUser user, String oldRefreshToken) {
+        validateRefreshToken(oldRefreshToken);
+        TokenDto newToken = tokenProvider.generateToken(user);
+        refreshTokenService.renewRefreshToken(user.getId(), newToken.getRefreshToken());
+        return newToken;
     }
 
     private void validateRefreshToken(String token) {
